+++ edulaw/pages/DynamicPOC/components/ChooseFormType.tsx
@@ -1,0 +1,42 @@
+import React, { ChangeEvent } from "react";
+import { Question, Answer, AnswersKeys, answers } from "../models";
+import { MyRadio } from './MyRadio';
+import { MyResult } from './MyResult';
+
+interface ChooseFormTypeProps {
+    question: Question,
+    onChange: (event: ChangeEvent<HTMLInputElement>) => void;
+}
+
+export const ChooseFormType: React.FC<ChooseFormTypeProps> = ({ ...props }): JSX.Element => {
+    let answerChoices: Answer[] = props.question.answers.map(function (answerId: number, i: number) {
+        const typedAnswerId = answerId.toString() as AnswersKeys;
+        return answers[typedAnswerId] as Answer;
+    })
+
+    switch (props.question.type) {
+        case 'RADIO': {
+            return <MyRadio
+                name={props.question.id.toString()}
+                label={props.question.question}
+                options={answerChoices}
+                onChange={props.onChange}
+            />
+        }
+        case 'RESULT': {
+            return <MyResult
+                question={props.question}
+            />
+        }
+        case 'TEXT': {
+            return <MyResult
+                question={props.question}
+            />
+        }
+        // TODO: To finish POC, finish implementing MyResult to handle that kind of answer?
+        // TODO: Other form types in general
+        default: {
+            return <div></div>
+        }
+    }
+}
