+++ edulaw/pages/DynamicPOC/components/MyResult.tsx
@@ -1,0 +1,11 @@
+import { Question } from "../models";
+
+interface MyResultProps {
+    question: Question
+}
+
+export const MyResult: React.FC<MyResultProps> = ({ ...props }): JSX.Element => {
+    return (
+        <div>{props.question.question}</div>
+    );
+}
