+++ edulaw/pages/DynamicPOC/constants/forms.json
@@ -1,0 +1,3 @@
+{
+    "animalForm": 0
+}
