+++ edulaw/pages/DynamicPOC/constants/index.ts
@@ -1,0 +1,3 @@
+export {default as forms} from './forms.json';
+export {default as questions} from './questions.json';
+export {default as answers} from './answers.json';
