--- edulaw/pages/DynamicPOC/formStore/animalForm.json
@@ -1,3 +1,0 @@
-{
-    "formStartingPoint": 0
-}
