+++ edulaw/pages/DynamicPOC/index.tsx
@@ -1,101 +1,17 @@
-import animalForm from './formStore/animalForm.json';
-import questions from './formStore/questions.json';
-import answers from './formStore/answers.json'; // TODO: Export these in directory
-
+import {Question, questions, QuestionsKeys, answers, AnswersKeys, forms} from './models';
 import { Form, Formik } from 'formik';
-import { MyRadio } from './MyRadio';
 import React, { ChangeEvent, useContext, useState } from 'react';
-import { FormAnswer, FormCtx, FormValues, SelectedAnswer } from './FormContext';
+import { FormAnswer, FormCtx, FormValues } from './FormContext'; // Note use of a different form context than before
+import { ChooseFormType } from './components/ChooseFormType';
 
-export interface Answer {
-    id: number,
-    type: string,
-    content: string,
-    route: number
-}
-
-type Answers = typeof answers;
-type AnswersKeys = keyof Answers;
-const typedAnswers = answers as Answers;
-
-type Questions = typeof questions;
-export type QuestionsKeys = keyof Questions;
-const typedQuestions = questions as Questions; // TODO: Do we use this / need it?
-
-
-interface Question {
-    id: number;
-    question: string,
-    description: string,
-    type: string,
-    answers: number[]
-}
-
-const firstQuestionId: QuestionsKeys = animalForm.formStartingPoint.toString() as QuestionsKeys;
-let startingQuestion: Question = typedQuestions[firstQuestionId] as Question;
+const firstQuestionId: QuestionsKeys = forms.animalForm.toString() as QuestionsKeys;
+let startingQuestion: Question = questions[firstQuestionId] as Question;
 let startingAnswer: FormAnswer;
 
-interface ChooseFormTypeProps {
-    question: Question,
-    onChange: (event: ChangeEvent<HTMLInputElement>) => void;
-}
-
-const ChooseFormType: React.FC<ChooseFormTypeProps> = ({ ...props }): JSX.Element => {
-    let answers: Answer[] = props.question.answers.map(function (answerId: number, i: number) {
-        const typedAnswerId = answerId.toString() as AnswersKeys;
-        return typedAnswers[typedAnswerId] as Answer;
-    })
-
-    console.log('CHOOSING FORM TIME');
-
-    switch (props.question.type) {
-        case 'RADIO': {
-            return <MyRadio
-                name={props.question.id.toString()}
-                label={props.question.question}
-                options={answers}
-                onChange={props.onChange}
-            />
-        }
-        case 'RESULT': {
-            return <MyResult
-                question={props.question}
-            />
-        }
-        case 'TEXT': {
-            return <MyResult
-                question={props.question}
-            />
-        }
-        // TODO: To finish POC, finish implementing MyResult to handle that kind of answer?
-        // TODO: Other form types in general
-        default: {
-            return <div></div>
-        }
-    }
-}
-
-interface MyResultProps {
-    question: Question
-}
-
-const MyResult: React.FC<MyResultProps> = ({ ...props }): JSX.Element => {
-    return (
-        <div>{props.question.question}</div>
-    );
-}
-
-const MyTextInput: React.FC<MyResultProps> = ({...props}): JSX.Element => {
-    return (
-        <div>{props.question.question}</div>
-    );
-}
-
 function getNextQuestion(answerId: AnswersKeys): Question {
-    let id: QuestionsKeys = typedAnswers[answerId].route.toString() as QuestionsKeys;
+    let id: QuestionsKeys = answers[answerId].route.toString() as QuestionsKeys;
     return questions[id] as Question;
 }
-
 
 const DynamicPOC: React.FC = () => {
     const { formValues, updateFormValues } = useContext(FormCtx);
@@ -122,8 +38,6 @@ const DynamicPOC: React.FC = () => {
         console.log(currentAnswer);
     }
 
-
-
     function _handleSubmit(values: FormValues) {
         // TODO: probably do something here, should probably break out into a handleNext and handleSubmit to handle the end of the form
         console.log(values);
@@ -144,7 +58,6 @@ const DynamicPOC: React.FC = () => {
                 }}
             >
                 <Form>
-                    {/* {_chooseFormType(currentQuestion)} */}
                     <ChooseFormType question={currentQuestion} onChange={_updateCurrentAnswer} />
                     <button type="submit">Submit</button> {/*TODO: Hide this at the end of the form */}
                 </Form>
