+++ edulaw/pages/DynamicPOC/models/form.ts
@@ -1,0 +1,8 @@
+import { forms } from '../constants';
+
+const typedForms = forms as Forms;
+
+export interface Forms {
+    animalForm: number;
+}
+export {typedForms as forms};
