+++ edulaw/pages/DynamicPOC/models/index.ts
@@ -1,0 +1,6 @@
+export { questions } from './question';
+export type { QuestionsKeys, Question } from './question';
+export { answers } from './answer';
+export type { AnswersKeys, Answer } from './answer';
+export { forms} from './form';
+export type { Forms } from './form';
